Save logs for some days

+++ src/Console/ApmClearCommand.php
@@ -37,13 +37,29 @@ class ApmClearCommand extends Command
     {
         $path = storage_path('app/apm/');
         $files = \File::allFiles($path);
+        $filesNamesToKeep = $this->getFilesNamesToKeep();
+
         foreach ($files as $file) {
             $filename = $file->getFilename();
             $pattern = '/^(apm-[0-9]{4}-[0-9]{2}-[0-9]{2}\.txt)$/';
             preg_match($pattern, $filename, $matches);
-            if (!empty($matches) && isset($matches[1])) {
+            if (!empty($matches) && isset($matches[1]) && ! in_array($filename, $filesNamesToKeep)) {
                 \File::delete($file);
             }
         }
+    }
+
+    private function getFilesNamesToKeep(): array
+    {
+        $keepForDays = config('apm.keep_for_days', 1);
+        $date = \Carbon\Carbon::toDay();
+        $filesNamesToKeep = [];
+
+        for ($i = 0; $i < $keepForDays; $i++) {
+            $filesNamesToKeep[] = 'apm-' . $date->toDateString() . '.txt';
+            $date->subDay();
+        }
+
+        return $filesNamesToKeep;
     }
 }
